:recycle: Refactoring: Alter class 구조 변경 (#6)
